refactor: rename to loading-bar.js
