+++ src/loading-bar.js
